feat: integrate classes to run full lighthouse audit
The logic in the index file is similar to the logic in the main.py
file in the lighthouse-docker repo, https://github.com/iFixit/lighthouse-docker/blob/master/lib/main.py.

The only difference is we will use the timestamp of the lighthouse report
to timestamp the metrics data sent to Datadog.