fix: add missing behavior to run both form factors
In lighthouse-docker we run both form factors, desktop and mobile.
To do this we need to import the associated config files for each
form factor.

We cannot just set the "formFactor" lh option to "desktop" or "mobile"
as there are other options that need to be set. For example, the
"throttling" option. These additional options were automatically set
when we ran lighthouse through the CLI tool in lighthouse-docker. However,
when we run lighthouse programmatically, we need to set these options
ourselves, or use the config files directly.