fix(chrome): use puppeteer instead of chrome-launcher
We need to use puppeteer instead of chrome-launcher because at the moment
some of the current Devs are having issues with chrome-launcher when
running in WSL. We don't have a docker image for vigilo yet, so we need
to use puppeteer for now.

We will not be passing a "Page" object to the lighthouse runner instead
of a "port" option. Aside from that, everything else should be the same.