fix: get puppeteer working in docker (M1 Mac)
Getting the puppeteer package to work in a docker container on an M1 Mac is a bit tricky. Besides the usual platform arm vs amd issues, there is also a problem with the dependencies
needed to run the browser. If the platform is not specified appropriately via `--platform=` or `export DOCKER_DEFAULT_PLATFORM=` then the vigilo will fail to build or run.

To get this working I exported the `DOCKER_DEFAULT_PLATFORM` variable to `linux/arm64` and then configured puppeteer to
use the linux installed version of chromium. This is done by setting the `PUPPETEER_SKIP_CHROMIUM_DOWNLOAD` environment variable to `true` and then setting the `PUPPETEER_EXECUTABLE_PATH` to the path of the installed chromium binary. At the same time, we need to override some
of the default puppeteer launch options to get it to work in the docker container. Mainly the `--no-sandbox` option is required since we are running as root in the container. Again these flags were taken from the `lighthouse-docker` project.