fix(Docker): set env to conditionally set hostname
When running the application via a docker container, the hostname is
set to the container id. This is not ideal in the long run because
we will constantly be creating/removing containers and the hostname
will change. All these container ids will then be stored in Datadog.

This sets an environment variable in the Docker image that will
conditionally set the hostname to just "docker-container" if the
we are running in a container. Otherwise, it will use the hostname
of the machine.