createDashboard: add page type widget requests
This adds the `requests` values to each pagetype widgets.
The `requests` property contains the most important part
of the dashboard which is the graph queries for our graphs.

Also, found out that the variables in the existing Datadog queries
were all lower case and used `_` instead of `-`. Thus, created
`formatDDQueryVar()` function to convert the audit names and page
types.