feat: add new method to DatadogClient to update metric metadata
This adds a new method to the DatadogClient class to update metric
metadata. One thing to note is there are two different MetricMetadata
types in the Datadog API. One is from v1 and the other is from v2. The
v2 MetricMetadata type is for the MetricSeries endpoint and is not
compatible with the v1 MetricMetadata type.