feat: integrate updating of metric metadata in main
When running the main script, we will now also update the metric metadata.
In order to do this, we need to retrieve some additional properties from
the audits in the lighthouse reports.

That said, the data points for the 'score' audit types do not adhere to
the same numeric unit type as the 'value' audit types. Therefore, their
unit field needs to be set to 'unit' according to the Datadog API docs.