refactor(createDashboard): update types and import new js file
By importing the new js file, the audits variable no longer had an any
type. However, because it was interpreted as any, the WidgetDefinition
was able to work with the creation of the dashboard. Now that the
audits variable is explicit type, everything broke. This fixes that by
updating the expected return types of the functions and using the
granular widget definitions instead of the generic one, i.e.
the GroupWidgetDefinition and TimeseriesWidgetDefinition.