feat: install tsup bundler
This bundler is a wrapper around esbuild, and allows us to easily
bundle our code for Node as ESM. In addition, it allows us to
better control over what we want to compile and will automatically
import dependent modules for the entry points we specify.

SWC, Rspack, Webpack, and vanilla esbuild were all considered, but
tsup was the easiest to use and configure for the simple use case of
compiling our ESM TS code for Node while still being in ESM format. On
top of that, it has finer control over the entrypoints and automatically
imports dependent modules.

The build.js file is a simple wrapper around tsup that will allow us
better control over the build process which will significant in later
commits.