feat: Add CLI options for tsup with yargs
This update introduces yargs options to selectively build 'core' or
'scripts' files. This enhancement allows for targeted builds, saving time
by avoiding unnecessary builds.

Tsup's automatic handling of dependencies eliminates the need for direct transpilation of core files for script builds and verbose pathing to core
and config files. This results in standalone output files, free from runtime external file dependencies.

Furthermore, the script files are now output to a 'bin' directory, maintaining a clean 'dist' directory and indicating that the bin files are standalone scripts executable from the CLI and are not part of the main app.