feat(build): add logic for non-bundling builds
For Docker builds, we don't want to bundle the config files into the
dist/index.js file. By bundling them, we lose the ability to change
them without rebuilding the image. Rebuilding the image every time
we want to change the config files is not ideal and would prevent us
from using the image for other sites.

Solving this problem is a bit tricky because we want to use the same
build script for both Docker and non-Docker builds, keep the independent
core and scripts bundling, and retain the path aliasing functionality.
To achieve all of this, I used the following:
- tsup's 'external' option to exclude the "@config" modules from the bundle
- The tsc-alias package to resolve the "@config" paths to a relative path
- The process.env.DOCKER variable to determine whether to use tsup's
  'external' option or not (this variable is set in the Dockerfile).

By doing all of this, we can now pass the config files into the
container as volumes and change them without rebuilding the image. At
the same time, we can still use the same build script for local
development and retain the simplicity of standalone builds.

The use of tsc-alias was discovered in this issue:
https://github.com/evanw/esbuild/issues/394#issuecomment-1537247216.
Without it, "@config/*" paths would remain in the bundle and would
cause errors as there is no way for "@config" to be resolved as
a local path.