fix(build): fix urls.json missing assertion
When we build the files for Docker, we will import the urls.json file
locally instead of the contents being bundled into the file. This
raises ERR_IMPORT_ASSERTION_TYPE_MISSING when we try to import the
file. To resolve this issue, we need to add the `type: "json"` to the
import statement as well as change the "module" and "moduleResolution"
options in the tsconfig.json file as import assertions are only supported
in 'esnext' or 'nodenext'.