fix(env): add comment to prevent wrapping of values with quotes
dotenv will automatically remove quotes from values, but if we pass
the .env file to the Docker container with the '--env-file' option,
the quotes will be preserved. This will cause 403 Forbidden errors with
the Datadog API because its interpreted the quotes as part of the
value. The simplest solution is to add a comment to the .env.template
and leave it up to the user to ensure they don't wrap values with
quotes.