fix(docker): add new entries to .dockerignore
We now build a bin directory, so we need to ignore it.

The config files, 'lh-config.js' and 'urls.json' will be volume mounted
into the container, so we can ignore them.

Lastly, we no longer want to include the .env file in the image, as
we will be deploying the image to ECR and we don't want our DD keys
in the image.