feat(helm): add lighthouse configmap and urls configmap
To run Vigilo, we need to have a `url.json` file and a `lh-config.js` file
within the container. This can be seen in `dev-scripts/run-docker.sh` file
when running the docker container directly. Therefore, in order to add
these files to the cronjob container we will volume mount them. To do that,
we need to create a configmap for each of these files to be mounted.

As such, this adds a configmap template for the `lh-config.js` file and
`urls.json` file which can then be modified in the `values.yaml` file.