feat(helm): add secret for datadog credentials
The last bit to getting the helm chart working is to add a secret for the
DataDog credentials (i.e. the .env file). This adds a secret template and
updates the cronjob to include the secret.

The creation of the secret can be turned off and we can point to a different
secret if needed.