refactor(core): only modify urls when capturing
Prior to this were were creating a new list of urls to audit and appending a
random parameter to each url. This is necessary to ensure that we are not
hitting a CDN cache when we are trying to audit the urls. However, when we send
the metrics to DataDog, we send the modified url. This causes the metrics to be
unique and harder to aggregate or create trends on.

We can't drop the group by 'url' in DataDog because we might be sending multiple
urls for the same page type which would cause the metrics to be aggregated
incorrectly.

Instead, let's only modify the urls when we are starting the lighthouse runner
which will allow us to send the original url to DataDog.