fix(createDashboard): fix the queries for the widgets
We've updated the lighthouse metrics to send both a 'score' and 'value' datapoint
for each audit. As such we need to update the queries to specify the new
subpath. For most of the audits, the 'value' datapoint is more useful than the
'score' datapoint. The only audits that we would use the 'score' datapoint are
those with null 'value' datapoints like the 'is_crawlable' audit.